test: move rss-fetch SSRF test to src/tests/security/ to match project convention
